update pyenv.sh to prevent bad PATH modifications
When pyenv user install is performed for user 1, if user 2 does not have access to user 1's home directory (which is the case on RHEL+clones) user 2 will still get the directory added to its path.

On login, `-bash: pyenv: command not found` will be displayed.

It's not too much of a problem (adding a nonexistent directory to PATH) but it's annoying to see. `-d` will prevent this behavior.